Create an install target for cmake
The public headers files are now in 'include/momemta', while the private
ones are in 'src/'. Users will only be able to include public files into
their code.

This is needed to allow MoMEMta to be used in other softwares.
