+++ core/include/LibraryManager.h
