+++ core/include/SharedLibrary.h
