+++ core/src/ModuleFactory.cc
@@ -17,7 +17,7 @@
  */
 
 
-#include <ModuleFactory.h>
+#include <momemta/ModuleFactory.h>
 
 template<> PluginFactory<ModuleFactory::type>& PluginFactory<ModuleFactory::type>::get() {
     static PluginFactory<ModuleFactory::type> s_instance;
