+++ core/src/Pool.cc
@@ -17,7 +17,7 @@
  */
 
 
-#include <Pool.h>
+#include <momemta/Pool.h>
 
 boost::any Pool::raw_get(const InputTag& tag) {
     auto it = m_pool.find(tag);
