+++ core/src/SharedLibrary.cc
