+++ core/src/logging.cc
