+++ include/momemta/Utils.h
@@ -24,7 +24,7 @@
 #include <vector>
 #include <algorithm>
 
-#include <Types.h>
+#include <momemta/Types.h>
 
 #define SQ(x) (x*x)
 #define CB(x) (x*x*x)
