+++ modules/Flatter.cc
@@ -17,8 +17,8 @@
  */
 
 
-#include <ConfigurationSet.h>
-#include <Module.h>
+#include <momemta/ConfigurationSet.h>
+#include <momemta/Module.h>
 
 class Flatter: public Module {
     public:
