+++ modules/GaussianTransferFunction.cc
@@ -19,10 +19,11 @@
 
 #include <Math/DistFunc.h>
 
-#include <ConfigurationSet.h>
-#include <Module.h>
-#include <Types.h>
-#include <Utils.h>
+#include <momemta/ConfigurationSet.h>
+#include <momemta/Module.h>
+#include <momemta/Types.h>
+#include <momemta/Utils.h>
+
 #include <logging.h>
 
 class GaussianTransferFunction: public Module {
